API finalizada, com exceção dos métodos DELETE

+++ src/main/java/br/ufal/ic/academico/DepartmentResources.java
@@ -1,8 +1,5 @@
 package br.ufal.ic.academico;
 
-import br.ufal.ic.academico.models.course.Course;
-import br.ufal.ic.academico.models.course.CourseDAO;
-import br.ufal.ic.academico.models.course.CourseDTO;
 import br.ufal.ic.academico.models.department.Department;
 import br.ufal.ic.academico.models.department.DepartmentDAO;
 import br.ufal.ic.academico.models.department.DepartmentDTO;
@@ -17,6 +14,7 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.ArrayList;
+import java.util.List;
 
 @Path("department")
 @Slf4j
@@ -25,14 +23,13 @@ import java.util.ArrayList;
 public class DepartmentResources {
     private final DepartmentDAO departmentDAO;
     private final SecretaryDAO secretaryDAO;
-    private final CourseDAO courseDAO;
 
     @GET
     @UnitOfWork
     public Response getAll() {
         log.info("getAll departments");
 
-        return Response.ok(departmentDAO.getAll()).build();
+        return Response.ok(departmentListToDTOList(departmentDAO.getAll())).build();
     }
 
     @POST
@@ -42,7 +39,7 @@ public class DepartmentResources {
         log.info("create department: {}", entity);
 
         Department d = new Department(entity);
-        return Response.ok(departmentDAO.persist(d)).build();
+        return Response.ok(new DepartmentDTO(departmentDAO.persist(d))).build();
     }
 
     @GET
@@ -51,7 +48,11 @@ public class DepartmentResources {
     public Response get(@PathParam("id") Long id) {
         log.info("get department: id={}", id);
 
-        return Response.ok(new DepartmentDTO(departmentDAO.get(id))).build();
+        Department d = departmentDAO.get(id);
+        if (d == null) {
+            return Response.status(404).entity("Department not found.").build();
+        }
+        return Response.ok(new DepartmentDTO(d)).build();
     }
 
     @PUT
@@ -62,8 +63,11 @@ public class DepartmentResources {
         log.info("update department: id={}", id);
 
         Department d = departmentDAO.get(id);
+        if (d == null) {
+            return Response.status(404).entity("Department not found.").build();
+        }
         d.update(entity);
-        return Response.ok(departmentDAO.persist(d)).build();
+        return Response.ok(new DepartmentDTO(departmentDAO.persist(d))).build();
     }
 
     @DELETE
@@ -73,6 +77,9 @@ public class DepartmentResources {
         log.info("delete department: id={}", id);
 
         Department d = departmentDAO.get(id);
+        if (d == null) {
+            return Response.status(404).entity("Department not found.").build();
+        }
         departmentDAO.delete(d);
         return Response.status(Response.Status.NO_CONTENT).build();
     }
@@ -83,13 +90,17 @@ public class DepartmentResources {
     public Response getAllSecretariesFromDepartment(@PathParam("id") Long id) {
         log.info("getAll secretaries from department {}", id);
 
-        ArrayList<Secretary> secretaries = new ArrayList<>();
         Department d = departmentDAO.get(id);
+        if (d == null) {
+            return Response.status(404).entity("Department not found.").build();
+        }
+
+        ArrayList<SecretaryDTO> secretaries = new ArrayList<>();
         if (d.getGraduation() != null) {
-            secretaries.add(d.getGraduation());
+            secretaries.add(new SecretaryDTO(d.getGraduation()));
         }
         if (d.getPostGraduation() != null) {
-            secretaries.add(d.getPostGraduation());
+            secretaries.add(new SecretaryDTO(d.getPostGraduation()));
         }
         return Response.ok(secretaries).build();
     }
@@ -101,10 +112,18 @@ public class DepartmentResources {
     public Response createSecretary(@PathParam("id") Long id, SecretaryDTO entity) {
         log.info("create secretary on department {}", id);
 
+        Department d = departmentDAO.get(id);
+        if (d == null) {
+            return Response.status(404).entity("Department not found.").build();
+        }
+
+        if (!entity.type.toUpperCase().equals("GRADUATION") && !entity.type.toUpperCase().equals("POST-GRADUATION")) {
+            return Response.status(400).entity("Invalid secretary type.").build();
+        }
+
         Secretary s = new Secretary(entity);
         secretaryDAO.persist(s);
 
-        Department d = departmentDAO.get(id);
         if (s.getType().equals("GRADUATION")) {
             if (d.getGraduation() != null) {
                 return Response.status(400).entity("Graduation Secretary already exists on this department.").build();
@@ -116,6 +135,14 @@ public class DepartmentResources {
             }
             d.setPostGraduation(s);
         }
-        return Response.ok(departmentDAO.persist(d)).build();
+        return Response.ok(new DepartmentDTO(departmentDAO.persist(d))).build();
+    }
+
+    private List<DepartmentDTO> departmentListToDTOList(List<Department> list) {
+        List<DepartmentDTO> dtoList = new ArrayList<>();
+        if (list != null) {
+            list.forEach(d -> dtoList.add(new DepartmentDTO(d)));
+        }
+        return dtoList;
     }
 }
